analyze/requirements.txt
numpy==1.15.0
redis==2.10.6
aiopg==0.15.0
psycopg2-binary==2.7.5
bcrypt==3.1.4
tornado==5.1
face_recognition